+++ APILambda/events/endEvent.ts
@@ -1,8 +1,8 @@
 import { conn } from "TiFBackendUtils"
+import { getEventSQL } from "TiFBackendUtils/TiFEventUtils"
 import { resp } from "TiFShared/api/Transport"
 import { failure, success } from "TiFShared/lib/Result"
 import { TiFAPIRouterExtension } from "../router"
-import { getEventSQL } from "../utils/eventDetails"
 
 /**
  * End or cancel an event given an event id.
