+++ APILambda/test/createCognitoUsers.ts
@@ -157,5 +157,5 @@ export const createCognitoTestAuthToken = async ( // TODO: Remove after we get c
   // @ts-ignore
   const { payload: { sub, name } } = jwt.decode(idToken, { complete: true })
 
-  return { auth: `Bearer ${idToken}`, id: sub, name, refreshAuth }
+  return { auth: `Bearer ${idToken}`, id: sub as string, name, refreshAuth }
 }
