+++ TiFBackendUtils/dateUtils.test.ts
@@ -1,4 +1,5 @@
-import { calcSecondsToStart, calcTodayOrTomorrow } from "./dateUtils"
+import dayjs from "dayjs"
+import { calcSecondsToStart, calcTodayOrTomorrow, isDayAfter } from "./dateUtils"
 import { todayTestDate, tomorrowTestDate } from "./test/dateHelpers"
 
 describe("calcSecondsToStart", () => {
@@ -26,5 +27,37 @@ describe("calcTodayOrTomorrow", () => {
 
   it("should return 'Tomorrow' if the secondsToStart is greater than or equal to SECONDS_IN_DAY", async () => {
     expect(calcTodayOrTomorrow(tomorrowTestDate())).toEqual("tomorrow")
+  })
+})
+
+describe("isDayAfter", () => {
+  beforeEach(() => {
+    jest.useFakeTimers()
+  })
+
+  afterEach(() => {
+    jest.useRealTimers()
+  })
+
+  it("should return false when endedDateTime is undefined", () => {
+    expect(isDayAfter()).toBe(false)
+  })
+
+  const fixedNow = new Date("2024-04-01T12:00:00Z")
+
+  it("should return false when less than 24 hours have passed since endedDateTime", () => {
+    jest.setSystemTime(fixedNow)
+
+    const endedDateTime = dayjs(fixedNow).subtract(23, "hour").toDate()
+
+    expect(isDayAfter(endedDateTime)).toBe(false)
+  })
+
+  it("should return true when at least 24 hours have passed since endedDateTime", () => {
+    jest.setSystemTime(fixedNow)
+
+    const endedDateTime = dayjs(fixedNow).subtract(24, "hour").toDate()
+
+    expect(isDayAfter(endedDateTime)).toBe(true)
   })
 })
